v1.2.0 - conditionally-compile extra implementations under submodule extra:: - to include in tests and benchmarks set RUSTFLAGS="--cfg extra" - replaced criterion with bencher and simplified benchmarks
