+++ .github/workflows/ci.yml
@@ -5,6 +5,8 @@ on:
 jobs:
   build:
     runs-on: ubuntu-latest
+    env:
+      RUSTFLAGS: --cfg extra
     steps:
     - uses: actions/checkout@v2
     - run: |
