+++ tests/test-shortscale.rs
@@ -1,8 +1,7 @@
 use shortscale;
 
-#[test]
-fn call_with_a_bunch_of_different_values() {
-    for (num, expected) in [
+const TESTS: [(u64, &str); 46] = [
+    /* 0 */
     (0, "zero"),
     (1, "one"),
     (10, "ten"),
@@ -13,6 +12,7 @@ fn call_with_a_bunch_of_different_values() {
     (30, "thirty"),
     (33, "thirty three"),
     (111, "one hundred and eleven"),
+    /* 0 */
     (120, "one hundred and twenty"),
     (121, "one hundred and twenty one"),
     (300, "three hundred"),
@@ -23,6 +23,7 @@ fn call_with_a_bunch_of_different_values() {
     (2_011, "two thousand and eleven"),
     (2_020, "two thousand and twenty"),
     (2_050, "two thousand and fifty"),
+    /* 0 */
     (2_300, "two thousand three hundred"),
     (2_301, "two thousand three hundred and one"),
     (30_020, "thirty thousand and twenty"),
@@ -33,6 +34,7 @@ fn call_with_a_bunch_of_different_values() {
     (1_000_000, "one million"),
     (1_001_000, "one million one thousand"),
     (1_002_000, "one million two thousand"),
+    /* 0 */
     (1_002_004, "one million two thousand and four"),
     (1_002_011, "one million two thousand and eleven"),
     (1_002_020, "one million two thousand and twenty"),
@@ -43,6 +45,7 @@ fn call_with_a_bunch_of_different_values() {
     (1_430_020, "one million four hundred and thirty thousand and twenty"),
     (1_430_920, "one million four hundred and thirty thousand nine hundred and twenty"),
     (1_999_001, "one million nine hundred and ninety nine thousand and one"),
+    /* 0 */
     (100_999_120, "one hundred million nine hundred and ninety nine thousand one hundred and twenty"),
     (999_999_120, "nine hundred and ninety nine million nine hundred and ninety nine thousand one hundred and twenty"),
     (420_000_999_015, "four hundred and twenty billion nine hundred and ninety nine thousand and fifteen"),
@@ -52,14 +55,34 @@ fn call_with_a_bunch_of_different_values() {
         nine hundred and ninety nine billion nine hundred and ninety nine million nine hundred and ninety nine thousand \
         nine hundred and ninety nine"),
     (1_999_999_999_999_999_999, "(big number)")
-    ].iter() {
+    ];
+
+#[test]
+fn test_shortscale() {
+    for (num, expected) in TESTS.iter() {
         println!("shortscale {}", num);
         assert_eq!(shortscale::shortscale(*num), String::from(*expected));
+    }
+}
+
+#[cfg(extra)]
+#[test]
+fn test_shortscale_extra() {
+    for (num, expected) in TESTS.iter() {
         println!("shortscale_string {}", num);
-        assert_eq!(shortscale::shortscale_vec_push(*num), String::from(*expected));
+        assert_eq!(
+            shortscale::extra::shortscale_vec_push(*num),
+            String::from(*expected)
+        );
         println!("shortscale_vec_concat {}", num);
-        assert_eq!(shortscale::shortscale_vec_concat(*num), String::from(*expected));
+        assert_eq!(
+            shortscale::extra::shortscale_vec_concat(*num),
+            String::from(*expected)
+        );
         println!("shortscale_string_join {}", num);
-        assert_eq!(shortscale::shortscale_string_join(*num), String::from(*expected));
+        assert_eq!(
+            shortscale::extra::shortscale_string_join(*num),
+            String::from(*expected)
+        );
     }
 }
